Factor out generic token-navigation routines from HTMLUtils
